UPDATE: rename custome.ts file
