+++ src/App.tsx
@@ -1,18 +1,20 @@
 import RestaurantList from "./components/RestaurantList"
-import { RestaurantType } from "./custome"
+import { RestaurantType } from "./custom"
 
 const restaurants: RestaurantType[] = [
   {
     id: "1",
     name: "Nandos",
     location: "Kampala Road",
-    status: "up"
+    status: "up",
+    path: "https://picsum.photos/200"
   },
   {
     id: "2",
     name: "KCF Kabalagala",
     location: "Kabalagala",
-    status: "up"
+    status: "up",
+    path: "https://picsum.photos/200"
   }
 ]
 
@@ -20,7 +22,7 @@ const restaurants: RestaurantType[] = [
 function App() {
   return (
     <>
-      <RestaurantList restaurants={restaurants}/>
+      <RestaurantList restaurants={restaurants} />
     </>
   )
 }
