+++ src/components/RestaurantList.tsx
@@ -1,4 +1,4 @@
-import { RestaurantType } from "../custome"
+import { RestaurantType } from "../custom"
 import RestaurantCard from "./RestaurantCard"
 
 type RestaurantsProps = {
@@ -6,13 +6,27 @@ type RestaurantsProps = {
 }
 
 const RestaurantList = ( { restaurants } : RestaurantsProps) => {
+
+    // const [restaurants, setRestaurants] =  useState<RestaurantType[]>([])
+
+    // useEffect(() => {
+    //     const fetchRestaurants = async(): Promise<RestaurantType[]> => {
+    //         const restaurants = await fetch('http://127.0.0.1:3000/restaurants')
+    //         .then(res => {
+    //             console.log(res.json())
+    //             return res.json()})
+    //         .catch(err => { if (err instanceof Error) console.log(err)})
+    //         return restaurants
+    //     }
+    //     fetchRestaurants().then(restaurants => setRestaurants(restaurants))
+    // })
+
     const restaurantCards = restaurants.map((restaurant: RestaurantType) =>
     <article key={restaurant.id}>
         <RestaurantCard 
-            id={restaurant.id}
             name={restaurant.name}
             location={restaurant.location}
-            status={restaurant.status}
+            path={restaurant.path}
         />
     </article>
     )
